Make the cedar device Bluetooth discoverable when awaiting a connection
Stellarium Plus on Android does not allow the user to select devices that are already bonded. It simply uses the BluetoothAdapter#startDiscovery() API to look for Bluetooth devices that are discoverable. 

SkySafari behaves correctly here, where it uses the android.bluetooth.devicepicker.action.LAUNCH intent action to have the system's Bluetooth device picker shown. That device picker includes devices that are already paired in addition to other devices that are discoverable.

The chances of Stellarium updating the behavior any time soon would be slim. So this is change is making the device visible/discoverable when there is no active connection. This isn't ideal as there is a small power hit when there is no connection, and it shouldn't be needed.

The workaround without this change is to start the pairing flow in Cedar Aim again, and while the device is pairable (for ~55 seconds) go to Stellarium and start the Bluetooth device selection flow.